Allow CI to pass on style warnings
This is a proposal to be less rigid with style requirements. Not all
observations that phpcs makes are intended by its authors and the
authors of standards like PSR-2/PSR-12 to be hard-and-fast rules that
cannot ever have exceptions. An example is characters per line limits,
which PSR-2 says this about

 > There MUST NOT be a hard limit on line length; the soft limit MUST
   be 120 characters; lines SHOULD be 80 characters or less.

IMO there is value in making warnings available for developers to
consider, (`composer phpcs`), but that this automation does not always
result in better code and contributors should be able to decide not to
comply with every observation phpcs is capable of making.